Resolved bugs
26 Feb 2025

Changed the width of the main components such that it is not taking up the full width of the browser and triggering the 'Resize Observer Loop' error.

Fixed the issue of the karaoke-style highlight that was not visually appearing when the stop button was clicked and then the play/pause button was clicked after.